% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absolute.r
\name{absolute}
\alias{absolute}
\title{Absolute value of a number}
\usage{
absolute(x1)
}
\arguments{
\item{x1}{number to be made positive}
}
\description{
Return the same magnitude, but always positive
}
\examples{
absolute(-13.26)
absolute(-999999)
}
