% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add3.r
\name{add3}
\alias{add3}
\title{An example 3 parts addition function}
\usage{
add3(n1, n2, n3)
}
\arguments{
\item{n1}{the first number to be added}

\item{n2}{the second number to be added}

\item{n3}{the third number to be added}
}
\description{
This function takes three numbers and combines them using addition
}
\examples{
add3(5, 6, 7)
}
