% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponentiate.r
\name{exponentiate}
\alias{exponentiate}
\title{Exponentiate an integer}
\usage{
exponentiate(base, power)
}
\arguments{
\item{base}{base to be exponentiated}

\item{power}{power of the exponentiation}
}
\description{
Provide an integer and then raise it to a real integer. Negative integers work, but not fractions.
}
\examples{
exponentiate(3, 2)
exponentiate(3, -2)
}
