% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial.r
\name{factorial}
\alias{factorial}
\title{An example factorial function}
\usage{
factorial(integer)
}
\arguments{
\item{integer}{A non-negative integer}
}
\value{
The factorial of the given integer
}
\description{
An example factorial function
}
\examples{
factorial(5)
}
