% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiply3.r
\name{multiply3}
\alias{multiply3}
\title{A function that multiplies three numbers}
\usage{
multiply3(a, b, c)
}
\arguments{
\item{a}{A numeric value}

\item{b}{A numeric value}

\item{c}{A numeric value}
}
\description{
This function takes three numeric inputs and multiplies them.
}
\examples{
multiply3(2, 3, 4)
}
