% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pctformat.r
\name{pctformat}
\alias{pctformat}
\title{a function to format numbers as percentages}
\usage{
pctformat(x, digits = 2)
}
\arguments{
\item{x}{A numeric value}

\item{digits}{An integer specifying the number of decimal places}
}
\description{
This function takes a number and formats it as a percentage.
}
\examples{
pctformat(0.1234) # returns "12.34\%"
}
