% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomOperation.R
\name{randomOperation}
\alias{randomOperation}
\title{Performs a random operation on the number by a random number between 1 and 1000}
\usage{
randomOperation(a)
}
\arguments{
\item{a}{an input number}
}
\description{
Performs a random operation on the number by a random number between 1 and 1000
}
