% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root.r
\name{root}
\alias{root}
\title{An example root function}
\usage{
root(num, n)
}
\arguments{
\item{num}{is the number to take the root of}

\item{x}{is the root degree (2 for square root, 3 for cubed root, etc.)}
}
\description{
An example root function
}
\examples{
root(25, 2)
}
