% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sin_squared_plus_cos_squared.r
\name{sin_squared_plus_cos_squared}
\alias{sin_squared_plus_cos_squared}
\title{An example sin squared plus cos squared function}
\usage{
sin_squared_plus_cos_squared(int)
}
\arguments{
\item{int}{A non-negative integer}
}
\value{
Calculates sin squared plus cos squared of the given integer. It should always return 1.
}
\description{
An example sin squared plus cos squared function
}
\examples{
sin_squared_plus_cos_squared(pi / 4)
}
