% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log10.r
\name{TheLog10}
\alias{TheLog10}
\title{An example log function of the log function}
\usage{
TheLog10(numb)
}
\arguments{
\item{numb}{the number to log}
}
\description{
This function does nothing useful but it is fun to create a function.
}
\examples{
log10(5)
}
