% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplydivide.r
\name{multiplydivide}
\alias{multiplydivide}
\title{multiplies 2 numbers and divides by a third}
\usage{
multiplydivide(m1, m2, d1)
}
\arguments{
\item{m1}{is the first number in the multiplication}

\item{m2}{is the second number in the multiplication}

\item{d1}{is the divisor in the function}
}
\description{
This function takes in 3 numbers, multiplies the first two and divides by the third
}
